Sleeve Generator
We asked and he delivered! Nettle Boy released a template for custom
sleeves today, so here's the implementation in this generator for it!
Source:
http://ilovebacons.com/threads/guide-to-re-animating-clothes-with-json.12019/page-9#post-97855

The PantsAndSleevesGenerator has been replaced with a SleeveGenerator,
so now you can drag sleeve sprite sheets on the sleeve generator and
outfit / pants sprite sheets on the pants generator.